+++ lib/ductwork/models/process.rb
@@ -11,6 +11,16 @@ module Ductwork
 
     validates :pid, uniqueness: { scope: :machine_identifier }
 
+    def self.adopt_or_create_current!
+      pid = ::Process.pid
+      machine_identifier = Ductwork::MachineIdentifier.fetch
+      last_heartbeat_at = Time.current
+
+      Ductwork::Process
+        .find_or_initialize_by(pid:, machine_identifier:)
+        .tap { |process| process.update!(last_heartbeat_at:) }
+    end
+
     def self.current
       pid = ::Process.pid
       machine_identifier = Ductwork::MachineIdentifier.fetch