+++ lib/ductwork/processes/pipeline_advancer_runner.rb
@@ -24,7 +24,7 @@ module Ductwork
       end
 
       def run
-        create_process!
+        adopt_or_create_process!
         start_pipeline_advancers
 
         Ductwork.logger.debug(
@@ -87,13 +87,9 @@ module Ductwork
         )
       end
 
-      def create_process!
+      def adopt_or_create_process!
         Ductwork.wrap_with_app_executor do
-          process = Ductwork::Process.find_or_initialize_by(
-            pid: ::Process.pid,
-            machine_identifier: Ductwork::MachineIdentifier.fetch
-          )
-          process.update!(last_heartbeat_at: Time.current)
+          Ductwork::Process.adopt_or_create_current!
         end
       end